refactor: move Agreement schema & interface to separate file
